fix: fixed GenericLoaderProps location
